Fix syntax highlighting
https://github.com/banga/git-split-diffs/commit/ba9ccc38e9ba33cb9b2125055fa16375aa2017f2
broke syntax highlighting. It specified the `langs` field to
`getHighlighter` which is great for speeding up start up time, but
requires loading languages manually, which is a bit fragile.

There were several issues with the previous implementation:
1. In order to load a language, we were only checking for lang.id in
   `BUNDLED_LANGUAGES`. We should've also looked at the `aliases`, since
   file extensions are usually a better match for those.
2. We were not `await`-ing the `loadLanguage` call, which meant that the
   `highlighter` call would break if the language wasn't already loaded
   (due to a previous loadLanguage invocation).
3. `loadLanguage` itself is buggy: https://github.com/shikijs/shiki/issues/438

I fixed the first two before finding the third one, so I've pushed the fixes to
a branch. For now, I'm reverting back to loading all languages to fix syntax
highlighting. I've added snapshot testing for syntax highlighting and some TODOs
to improve things in the future.