Handle characters that take multiple columns
Some characters, such as 한 are displayed using multiple characters in
the terminal. https://en.wikipedia.org/wiki/Duospaced_font and
https://en.wikipedia.org/wiki/Halfwidth_and_fullwidth_forms describe
their behavior. It's possible that the behavior is different if you use
non-monospace fonts, but given that we can't detect fonts, making the
behavior correct for monospace fonts makes the most sense.