Fix occasional trailing `\r` in git output
For e.g. when displaying src/compiler/checker.ts in
https://github.com/microsoft/TypeScript/commit/6823559fa9730085d85ba838b9e36a0e1c311ceb,
there were trailing `\r`s in some of the lines, which showed up as ^M in
the terminal and an extra newline.

The cause turns out to be intervening ANSI color codes between \r and
\n. It would be easier to strip out color codes from the input earlier,
but we preserve it to avoid changing the output of other git commands
like `git log --graph`.