Incomplete migration to ESM
Upgrading the version for `terminal-size`, which forces us to switch to
ESM and bump the min node version to 18.

I got the build to work with esbuild using --format=esm. Also realized
that we can remove the `bin/git-split-diffs` file and instead directly
specify `index.mjs` in the `bin` field in `package.json` after inserting
a shebang line at the top using esbuild.

The stuff that doesn't work is:
- `yarn benchmark`, likely because the `benchmark` lib generates the
  module in weird way that conversion to ESM doesn't work. I should find
  a more modern benchmarking lib.
- `yarn test` fails because jest and ts-jest's ESM support seems kind of
  finicky.