Implement unified diff rendering for combined diffs
The naming is getting confusing here, but the approach we had for
rendering a unified diff for "unified diffs" wasn't great for "combined
diffs" because it would end up repeating additions/deletions.

So, I added a separate implementation which essentially renders what you
would see in the last column of the split diff. This required showing
deletions in the split diff as well, which I had previously skipped. I
think that still looks fine, so will let it be.