Handle combined diffs (#51)
This adds support for "combined diffs", which is the default format in which git displays merge commits. There wasn't one obvious approach to display them in split and unified formats, so I've chosen what looked reasonable to me. Github itself appears to show the output of git diff --first-parent instead.