Tenative fix for #16
A theory from claude is that since we were fire-and-forget-ing `main`,
the node process can die before flushing the full contents to stdout,
leaving `less` in a different mode because of control characters.

I'm not completely convinced because I/O should keep the process alive,
but maybe there's still a race condition possible and this change seems
harmless.